+++ src/helpers/errors.ts
@@ -1,0 +1,8 @@
+// TODO custom validation error
+
+export class ValidationError extends Error {
+  constructor(message: string) {
+    super(message);
+    Object.setPrototypeOf(this, ValidationError.prototype);
+  }
+}