+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 - added waitForConnectionStateChange
 - added connect() which returns a promise
 - changed disconnect() to include code and reason. Returns a promise now
+
+# 0.9.15
+
+- added Typebox validator
+- improved validation error reporting
+- moved reconnecting-websocket to peer dependency